Rename repo and application to nextflow-runner-service
