+++ .gitignore
@@ -1,6 +1,6 @@
 # Files autogenerated when making docs
-docs/sequencing_report_service.*.rst
-docs/sequencing_report_service.rst
+docs/nextflow_runner_service.*.rst
+docs/nextflow_runner_service.rst
 docs/modules.rst
 docs/routes.rst
