+++ README.md
@@ -1,7 +1,7 @@
 Sequencing Report Service
 =========================
 
-Service used to start nextflow pipelines with passed config variables 
+Service used to start nextflow pipelines with passed config variables
 
 How to configure pipelines
 --------------------------
@@ -78,7 +78,7 @@ input_samplesheet_content: |
 ```
 
 
-Installing sequencing-report-service
+Installing nextflow-runner-service
 ----------------
 1. Clone the repo
 
@@ -129,7 +129,7 @@ uv remove <'package'>  # Removes dependencies in the project and removes in the 
 #### Dependency Locking
 
 We use UV's lockfile functionality to ensure reproducible builds:
-NOTE: `uv add` and `uv remove` usually edit both pyproject.toml and uv.lock but 
+NOTE: `uv add` and `uv remove` usually edit both pyproject.toml and uv.lock but
 one can run the command below if you delete the lock file to regenerate
 
 1. Generate/update lockfile:
@@ -137,4 +137,4 @@ one can run the command below if you delete the lock file to regenerate
 uv lock
 ```
 
-Sequencing-report-service project *version* is documented in the ```pyproject.toml``` file and should be updated there to match the releases.
+nextflow-runner-service project *version* is documented in the ```pyproject.toml``` file and should be updated there to match the releases.
