+++ alembic/env.py
@@ -7,7 +7,7 @@ from alembic import context
 from sqlalchemy import engine_from_config, pool
 from logging.config import dictConfig
 
-from sequencing_report_service.models.db_models import SQLAlchemyBase
+from nextflow_runner_service.models.db_models import SQLAlchemyBase
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
