+++ config/logger.config
@@ -18,7 +18,7 @@ handlers:
         class: logging.handlers.RotatingFileHandler
         level: DEBUG  # By default set to WARNING, but can be changed on runtime
         formatter: simple
-        filename: sequencing-report-service.log
+        filename: nextflow-runner-service.log
         maxBytes: 10485760  # 10MB
         backupCount: 20
         encoding: utf8
