+++ nextflow_runner_service/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-"""Top-level package for sequencing-report-service."""
+"""Top-level package for nextflow-runner-service."""
 
 __author__ = """SNP&SEQ Technology Platform, National Genomics Infrastructure, Uppsala University"""
 __email__ = 'dataoperations@medsci.uu.se'
