+++ nextflow_runner_service/handlers/job_handler.py
@@ -10,11 +10,11 @@ from tornado.web import HTTPError
 
 from arteria.web.handlers import BaseRestHandler
 
-from sequencing_report_service.handlers import ACCEPTED, NOT_FOUND, FORBIDDEN
-from sequencing_report_service.exceptions import UnableToStopJob, RunfolderNotFound
+from nextflow_runner_service.handlers import ACCEPTED, NOT_FOUND, FORBIDDEN
+from nextflow_runner_service.exceptions import UnableToStopJob, RunfolderNotFound
 import importlib.metadata
 
-version = importlib.metadata.version("sequencing-report-service")
+version = importlib.metadata.version("nextflow-runner-service")
 
 
 class OneJobHandler(BaseRestHandler):
@@ -126,7 +126,7 @@ class JobStartHandler(BaseRestHandler):
         try:
             request_data = self.body_as_object()
             runfolder_path = self.runfolder_repo.get_runfolder(runfolder)
-            
+
             job_id = self.runner_service.start(
                 pipeline,
                 runfolder_path=runfolder_path,
